# backend/api/docs/swagger.yaml
basePath: /
definitions:
  echo-backend_internal_models_auth.ApiKey:
    properties:
      created_at:
        example: "2026-01-15T10:30:00Z"
        type: string
      id:
        example: key_a1b2c3d4e5f6
        type: string
      name:
        example: Production API Key
        type: string
      prefix:
        example: sk_a1b2c3d4
        type: string
      scopes:
        example:
        - read
        - write
        - admin
        items:
          type: string
        type: array
      status:
        example: active
        type: string
      user_id:
        example: "1"
        type: string
    type: object
  echo-backend_internal_models_auth.User:
    properties:
      created_at:
        example: "2026-01-15T10:30:00Z"
        type: string
      email:
        example: jane@example.com
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Jane Doe
        type: string
      role:
        example: user
        type: string
      updated_at:
        example: "2026-01-15T10:30:00Z"
        type: string
    type: object
...
  internal_handler_admin.createAPIKeyRequest:
    properties:
      name:
        example: Production Key
        type: string
      scopes:
        example:
        - read
        - write
        items:
          type: string
        type: array
    type: object
  internal_handler_auth.loginRequest:
    properties:
      email:
        example: jane@example.com
        type: string
      password:
        example: P@ssw0rd!23
        type: string
    type: object
  internal_handler_auth.registerRequest:
    properties:
      email:
        example: jane@example.com
        type: string
      name:
        example: Jane Doe
        type: string
      password:
        example: P@ssw0rd!23
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@echo.app
    name: Echo API Support
    url: https://echo.app/support
  description: Core API Server for Echo Platform providing Auth, Chat, Sessions, Models,
    Settings, Admin, Memory, and System Services.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://echo.app/terms
  title: Echo Backend API
  version: "1.0"
paths:
  /api/v1/admin/api-keys:
    get:
      description: Returns all API keys (admin)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/echo-backend_internal_models_auth.ApiKey'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List API keys
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Generates and stores a new API key (admin)
      parameters:
      - description: API key payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler_admin.createAPIKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create an API key
      tags:
      - Admin
  /api/v1/admin/api-keys/{id}:
    delete:
      description: Revokes an API key by ID (admin)
      parameters:
      - description: API key ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Revoke an API key
      tags:
      - Admin
  /api/v1/admin/stats:
    get:
      description: Returns total and active API key counts (admin)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get API key statistics
      tags:
      - Admin
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT token
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler_auth.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login with email and password
      tags:
      - Auth
  /api/v1/auth/logout:
    post:
      description: Clears the authentication cookie
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - Auth
  /api/v1/auth/me:
    get:
      description: Returns the authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/echo-backend_internal_models_auth.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - Auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account and returns a JWT token
      parameters:
      - description: Registration payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_handler_auth.registerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token authorization header (Format: Bearer <JWT>).'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"